@@ -6,26 +6,33 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.18.2 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-
+  flutter_localizations:
+    sdk: flutter
+  intl: ^0.18.0
   cupertino_icons: ^1.0.5
-  widget_with_codeview: ^2.0.1+2
+  widget_with_codeview: ^3.0.1
   page_transition: ^2.0.9
   url_launcher: ^6.1.6
   shared_preferences: ^2.0.15
   dart_code_viewer2: ^1.0.0
   font_awesome_flutter: ^10.2.1
 
+dependency_overrides:
+  float_column: ^2.0.1
+  google_fonts: ^5.1.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
   flutter_lints: ^2.0.1
-  flutter_launcher_icons: ^0.10.0
+  flutter_launcher_icons: ^0.13.1
+  import_sorter: ^4.6.0
 
 flutter_icons:
   android: "launcher_icon"
@@ -43,9 +50,14 @@ flutter_icons:
     image_path: "assets/icons/app_icon.png"
     icon_size: 48 # min:48, max:256, default: 48
 
+  import_sorter:
+    comments: false
+#  flutter pub run import_sorter:main
+
 flutter:
 
   uses-material-design: true
+  generate: true
 
   assets:
     - assets/images/
