@@ -1,3 +1,4 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
 
 class WrapView extends StatelessWidget {
@@ -5,13 +6,13 @@ class WrapView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return const Scaffold(
       body: Center(
         /// A widget that displays its children in multiple horizontal or vertical runs.
         child: Wrap(
           spacing: 8.0, // gap between adjacent chips
           runSpacing: 4.0, // gap between lines
-          children: const<Widget>[
+          children: <Widget>[
             Text('Hamilton'),
             Text('Lafayette'),
             Text('Mulligan'),
